backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.6.2=pyhd8ed1ab_0
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.10.11=hca03da5_0
certifi=2022.12.7=py310hca03da5_0
cffi=1.15.1=py310h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
cryptography=38.0.4=py310h834c97f_0
dnspython=2.2.1=py310hca03da5_0
email-validator=1.3.0=pyhd8ed1ab_0
fastapi=0.89.0=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
gnutls=3.6.15=h887c41c_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.0.0=py310hca03da5_3
hpack=4.0.0=py_0
httpcore=0.16.3=pyhd8ed1ab_0
httptools=0.5.0=py310h8e9501a_1
httpx=0.23.3=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd3eb1b0_0
icu=68.1=hc377ac9_0
idna=3.4=py310hca03da5_0
itsdangerous=2.1.2=pyhd8ed1ab_0
jinja2=3.1.2=py310hca03da5_0
jpeg=9e=h1a28f6b_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libcxx=14.0.6=h848a8c0_0
libdeflate=1.8=h1a28f6b_5
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
liblapack=3.9.0=16_osxarm64_openblas
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.37=hb8d0fd4_0
libsqlite=3.40.0=h76d750c_0
libtasn1=4.16.0=h1a28f6b_0
libtiff=4.5.0=h2fd578a_0
libunistring=0.9.10=h1a28f6b_0
libuv=1.44.2=he4db4b2_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=h68602c7_0
libwebp-base=1.2.4=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
libzlib=1.2.13=h03a7124_4
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py310h1a28f6b_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=pypi_0
ncurses=6.3=h1a28f6b_3
nettle=3.7.3=h84b5d62_1
numpy=1.24.1=py310h5d7c261_0
openh264=1.8.0=h98b2900_0
openssl=1.1.1s=h1a28f6b_0
pillow=9.3.0=py310hf4a492f_1
pip=22.3.1=py310hca03da5_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.4=py310h8e9501a_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.8=hf452327_0_cpython
python-dotenv=0.21.0=py310hca03da5_0
python-multipart=0.0.5=py_0
python_abi=3.10=3_cp310
pytorch=2.0.0.dev20230113=py3.10_0
pyyaml=6.0=py310h80987f9_1
readline=8.2=h1a28f6b_0
requests=2.28.1=py310hca03da5_0
rfc3986=1.5.0=pyhd8ed1ab_0
setuptools=65.6.3=py310hca03da5_0
six=1.15.0=pyhd3eb1b0_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlite=3.40.1=h7a7dc30_0
starlette=0.22.0=pyhd8ed1ab_0
sympy=1.11.1=py310hca03da5_0
tk=8.6.12=hb8d0fd4_0
torchaudio=2.0.0.dev20230113=py310_cpu
torchvision=0.15.0.dev20230113=py310_cpu
typing-extensions=4.4.0=py310hca03da5_0
typing_extensions=4.4.0=py310hca03da5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py310hca03da5_0
uvicorn=0.20.0=py310hbe9552e_1
uvloop=0.17.0=py310h8e9501a_1
watchfiles=0.18.1=py310h1dc5ec9_0
websockets=10.4=py310h8e9501a_1
wheel=0.36.2=pyhd3eb1b0_0
x264=1!152.20180806=h1a28f6b_0
xz=5.2.8=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zlib=1.2.13=h03a7124_4
zstd=1.5.2=h8574219_0
